+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Added
 - Parsing of alternative PHPDoc array types (array shapes, object-like arrays) as mixed[]
 
+### Fixed
+- Parsing of compound typed array types, e.g. (int[]|string)[]
+
 ## 0.12.4 - 2019-11-21
 ### Fixed
 - Useless PHPDoc for const is no longer reported when doc type is incomplete (e.g. array)