+++ CHANGELOG.md
@@ -4,15 +4,13 @@ All notable changes to `phpcs-type-sniff` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## Next
+## 0.13.0 - 2019-12-09
 ### Removed
 - Setting `FqcnMethodSniff.usefulTags`
-
 ### Added
 - Parsing of alternative PHPDoc array types (array shapes, object-like arrays) as mixed[]
 - Setting `FqcnMethodSniff.invalidTags`. Specifies which methods tags should be removed.
 - Setting `FqcnMethodSniff.reportMissingTags`.
-
 ### Fixed
 - Parsing of compound typed array types, e.g. (int[]|string)[]
 - Parsing of generic tags with parentheses, no longer part of tag name, Dynamic content is now tag content.